LogpointApi.IncidentApi (logpoint_api v0.2.3)

This module provides an implementation of the Logpoint Incident API.

Functions

add_comments(ip, credential, incident_comment_data)

@spec add_comments(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.IncidentApi.IncidentCommentData.t()
) :: {:ok, map()} | {:error, String.t()}

Add comments to a list of incidents.

assign_incidents(ip, credential, incident_ids, assignee_id)

@spec assign_incidents(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.IncidentApi.IncidentIDs.t(),
  String.t()
) :: {:ok, map()} | {:error, String.t()}

Assign or re-assign a list of incidents.

close_incidents(ip, credential, incident_ids)

@spec close_incidents(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.IncidentApi.IncidentIDs.t()
) ::
  {:ok, map()} | {:error, String.t()}

Close a list of incidents.

get_data_from_incident(ip, credential, incident)

@spec get_data_from_incident(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.IncidentApi.Incident.t()
) :: {:ok, map()} | {:error, String.t()}

Get a specific incident and its related data.

get_incident_states(ip, credential, time_range)

@spec get_incident_states(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.IncidentApi.TimeRange.t()
) ::
  {:ok, map()} | {:error, String.t()}

Get the states of incidents within a specific time range.

get_incidents(ip, credential, time_range)

@spec get_incidents(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.IncidentApi.TimeRange.t()
) ::
  {:ok, map()} | {:error, String.t()}

Get all incidents within a given time range.

get_users(ip, credential)

@spec get_users(String.t(), LogpointApi.Credential.t()) ::
  {:ok, map()} | {:error, String.t()}

Get users.

reopen_incidents(ip, credential, incident_ids)

@spec reopen_incidents(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.IncidentApi.IncidentIDs.t()
) ::
  {:ok, map()} | {:error, String.t()}

Reopen a list of incidents.

resolve_incidents(ip, credential, incident_ids)

@spec resolve_incidents(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.IncidentApi.IncidentIDs.t()
) ::
  {:ok, map()} | {:error, String.t()}

Resolve a list of incidents.