LogpointApi.SearchApi (logpoint_api v0.2.3)

This module provides an implementation of the Logpoint search API.

Functions

get_devices(ip, credential)

@spec get_devices(String.t(), LogpointApi.Credential.t()) ::
  {:ok, map()} | {:error, String.t()}

Get the devices of the instance.

get_livesearches(ip, credential)

@spec get_livesearches(String.t(), LogpointApi.Credential.t()) ::
  {:ok, map()} | {:error, String.t()}

Get all live search on a machine.

get_logpoints(ip, credential)

@spec get_logpoints(String.t(), LogpointApi.Credential.t()) ::
  {:ok, map()} | {:error, String.t()}

Get all logpoints of a Logpoint instance.

get_repos(ip, credential)

@spec get_repos(String.t(), LogpointApi.Credential.t()) ::
  {:ok, map()} | {:error, String.t()}

Get the repos of an instance.

get_search_id(ip, credential, query)

@spec get_search_id(
  String.t(),
  LogpointApi.Credential.t(),
  LogpointApi.SearchApi.Query.t()
) ::
  {:ok, map()} | {:error, String.t()}

Create a search and get its search id.

get_search_result(ip, credential, search_id)

@spec get_search_result(String.t(), LogpointApi.Credential.t(), String.t()) ::
  {:ok, map()} | {:error, String.t()}

Retrieve the search result of a specific search id.

get_user_timezone(ip, credential)

@spec get_user_timezone(String.t(), LogpointApi.Credential.t()) ::
  {:ok, map()} | {:error, String.t()}

Get the current users time zone information.